ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq==0.13.1
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
